Setup Folder Structure according to Figma Chart
