+++ src/components/Carousel/Carousel.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Carousel = () => {
+  return <h1>A Carousel</h1>
+} 
+
+export default Carousel;
