+++ src/components/Home/Home.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Nav from '../../containers/Nav/Nav'
+
+
+const Home = () => {
+  return (
+      <Nav />
+  )
+}
+export default Home;
